Version member in pci-bios structure is an unsigned long not a short Contributed by David Galvez.

+++ sys/pcibios.c
@@ -91,7 +91,7 @@ static PCIBIOS_MINT pcibios_MiNT = { {
 # endif
 
 
-ushort pcibios_installed = 0;
+ulong pcibios_installed = 0;
 
 PCIBIOS *pcibios = NULL;
 void *tab_funcs_pci;
